Renamed tester.py file
