+++ test/sband/test_sband.py
@@ -25,7 +25,7 @@ import sys
 from os import path
 sys.path.append("./test")
 
-from tester import Tester
+from groundstation_tester import Tester
 
 tester = Tester() #call to initialize local test class
