+++ test/test_dfgm.py
@@ -26,7 +26,7 @@ import time
 from os import path
 sys.path.append("./test")
 
-from tester import Tester
+from groundstation_tester import Tester
 
 tester = Tester() #call to initialize local test class
