+++ test/test_eps_hk_only.py
@@ -22,7 +22,7 @@
 
 import time
 import numpy as np
-from tester import Tester
+from groundstation_tester import Tester
 
 tester = Tester()  # call to initialize local test class
