+++ test/test_flatsat_adcs.py
@@ -23,7 +23,7 @@ import time
 import sys
 from os import path
 
-from tester import Tester
+from groundstation_tester import Tester
 from eps.expected_eps_hk import expected_EPS_HK
 from test_full_hk import testSystemWideHK
