+++ test/test_flatsat_aurora.py
@@ -24,7 +24,7 @@ import numpy as np
 import sys
 import os
 
-from tester import Tester
+from groundstation_tester import Tester
 from eps.expected_eps_hk import expected_EPS_HK
 from test_full_hk import testSystemWideHK
