+++ test/test_flatsat_deployables.py
@@ -22,7 +22,7 @@
 import time
 import numpy as np
 
-from tester import Tester
+from groundstation_tester import Tester
 from eps.expected_eps_hk import expected_EPS_HK
 from test_full_hk import testSystemWideHK
