+++ test/uhf/test_uhf_hardware.py
@@ -27,7 +27,7 @@ import sys
 from os import path
 sys.path.append("./test")
 
-from tester import Tester
+from groundstation_tester import Tester
 
 tester = Tester() #call to initialize local test class
